+++ src/components/FeatureCard.jsx
@@ -2,15 +2,17 @@ import React from "react";
 
 const FeatureCard = ({ title, description, image, active, onToggle }) => {
   return (
-    <div className="relative rounded-xl p-[2px] bg-gradient-to-b from-blue-500 via-blue-300 to-white">
-      <div className="bg-gradient-to-br from-[#06172d] to-[#0e2c4f] rounded-xl p-4 h-full text-white">
+    <div className="relative rounded-xl p-[1px] bg-gradient-to-b from-blue-800 via-blue-300 to-gray-300">
+    
+
+      <div className="bg-gradient-to-br from-[#06172d] to-[#0e2c4f] rounded-xl p-4 h-full  text-white">
         <img
           src={image}
           alt={title}
           className="w-full h-40 object-cover rounded-md mb-3"
         />
-        <h3 className="text-lg font-semibold mb-1">{title}</h3>
-        <p className="text-sm text-gray-300 mb-4">{description}</p>
+        <h3 className="text-md  mb-1">{title}</h3>
+        <p className="text-sm text-gray-300 pt-2 mb-4">{description}</p>
 
         {/* Toggle Switch */}
         <label className="inline-flex items-center cursor-pointer">
